Indexing refactor:
* create splitgraph.core.indexing package
* extract current (range) indexing into a separate module in it (+ its relevant methods)
* move bloom there as well
* simplify/remove a couple of unneeded methods in FragmentManager
* idea is to make it easier to add pluggable indexes in the future with a consistent interface:
  * generate_index takes engine, object ID(s?), changeset, column + extra index params (e.g. probability for bloom) and returns an index "structure" to add to the metadata
  * filter_index takes object IDs and qualifiers and does filtering (either by loading the index like bloom does or querying it directly in Postgres).
